+++ onadata/apps/api/tests/viewsets/test_data_viewset.py
@@ -32,13 +32,14 @@ from onadata.apps.main import tests as main_tests
 from onadata.apps.main.models import UserProfile
 from onadata.apps.main.models.meta_data import MetaData
 from onadata.apps.main.tests.test_base import TestBase
+from onadata.apps.messaging.constants import XFORM, SUBMISSION_DELETED
 from onadata.libs import permissions as role
 from onadata.libs.permissions import ReadOnlyRole, EditorRole, \
     EditorMinorRole, DataEntryOnlyRole, DataEntryMinorRole
-from onadata.libs.utils.common_tags import MONGO_STRFTIME
-from onadata.libs.utils.logger_tools import create_instance
 from onadata.libs.serializers.submission_review_serializer import \
     SubmissionReviewSerializer
+from onadata.libs.utils.common_tags import MONGO_STRFTIME
+from onadata.libs.utils.logger_tools import create_instance
 
 
 @urlmatch(netloc=r'(.*\.)?enketo\.ona\.io$')
@@ -90,7 +91,9 @@ class TestDataViewSet(TestBase):
         self.extra = {
             'HTTP_AUTHORIZATION': 'Token %s' % self.user.auth_token}
 
-    def test_data(self):
+    @patch('onadata.apps.logger.models.instance.send_message')
+    def test_data(self, mock_send_message):
+        """Test DataViewSet list"""
         self._make_submissions()
         view = DataViewSet.as_view({'get': 'list'})
         request = self.factory.get('/', **self.extra)
@@ -122,6 +125,10 @@ class TestDataViewSet(TestBase):
         self.assertNotEqual(response.get('Cache-Control'), None)
         self.assertIsInstance(response.data, dict)
         self.assertDictContainsSubset(data, response.data)
+        self.assertTrue(mock_send_message.called)
+        mock_send_message.called_with(
+            dataid, formid, XFORM,
+            request.user, SUBMISSION_DELETED)
 
     @override_settings(STREAM_DATA=True)
     def test_data_streaming(self):
@@ -1449,7 +1456,10 @@ class TestDataViewSet(TestBase):
         instance.save()
         self.assertEqual(mock.call_count, 1)
 
-    def test_deletion_of_bulk_submissions(self):
+    @patch(
+        'onadata.apps.api.viewsets.data_viewset.send_message')
+    def test_deletion_of_bulk_submissions(self, send_message_mock):
+
         self._make_submissions()
         self.xform.refresh_from_db()
         formid = self.xform.pk
@@ -1488,6 +1498,10 @@ class TestDataViewSet(TestBase):
             response.data.get('message'),
             "%d records were deleted" % len(records_to_be_deleted)
         )
+        self.assertTrue(send_message_mock.called)
+        send_message_mock.called_with(
+            [str(i.pk) for i in records_to_be_deleted], formid, XFORM,
+            request.user, SUBMISSION_DELETED)
         self.xform.refresh_from_db()
         current_count = self.xform.instances.filter(deleted_at=None).count()
         self.assertNotEqual(current_count, initial_count)
@@ -2304,7 +2318,8 @@ class TestDataViewSet(TestBase):
             "status": SubmissionReview.APPROVED
         }
 
-        serializer_instance = SubmissionReviewSerializer(data=data)
+        serializer_instance = SubmissionReviewSerializer(data=data, context={
+            "request": request})
         serializer_instance.is_valid()
         serializer_instance.save()
         instance.refresh_from_db()
@@ -2321,7 +2336,8 @@ class TestDataViewSet(TestBase):
             "status": SubmissionReview.PENDING
         }
 
-        serializer_instance = SubmissionReviewSerializer(data=data)
+        serializer_instance = SubmissionReviewSerializer(data=data, context={
+            "request": request})
         serializer_instance.is_valid()
         serializer_instance.save()
         instance.refresh_from_db()
@@ -2339,7 +2355,8 @@ class TestDataViewSet(TestBase):
             "note": "Testing"
         }
 
-        serializer_instance = SubmissionReviewSerializer(data=data)
+        serializer_instance = SubmissionReviewSerializer(data=data, context={
+            "request": request})
         serializer_instance.is_valid()
         serializer_instance.save()
         instance.refresh_from_db()