+++ onadata/apps/api/tests/viewsets/test_openid_connect_viewset.py
@@ -1,6 +1,7 @@
 """
 Test OpenIDViewset module
 """
+from django.contrib.auth.models import User
 from django.test.utils import override_settings
 
 from mock import patch
@@ -137,9 +138,37 @@ class TestOpenIDConnectViewSet(TestAbstractViewSet):
             'email': 'john@doe.com'
         }
         data = {'id_token': 124, 'username': 'john'}
+        user_count = User.objects.filter(username='john').count()
         request = self.factory.post('/', data=data)
         response = self.view(request, openid_connect_provider='msft')
+        self.assertEqual(
+            user_count + 1, User.objects.filter(username='john').count())
         self.assertEqual(response.status_code, 302)
+
+        # Uses last_name as first_name if missing
+        mock_get_decoded_id_token.return_value = {
+            'family_name': 'davis',
+            'email': 'davis@justdavis.com'
+        }
+        data = {'id_token': 124, 'username': 'davis'}
+        user_count = User.objects.filter(username='davis').count()
+        request = self.factory.post('/', data=data)
+        response = self.view(request, openid_connect_provider='msft')
+        self.assertEqual(
+            user_count + 1, User.objects.filter(username='john').count())
+        self.assertEqual(response.status_code, 302)
+        user = User.objects.get(username='davis')
+        self.assertEqual(user.first_name, 'davis')
+
+        # Returns a 400 response if both family_name and given_name
+        # are missing
+        mock_get_decoded_id_token.return_value = {
+            'email': 'jake@doe.com'
+        }
+        data = {'id_token': 124, 'username': 'jake'}
+        request = self.factory.post('/', data=data)
+        response = self.view(request, openid_connect_provider='msft')
+        self.assertEqual(response.status_code, 400)
 
     @override_settings(OPENID_CONNECT_PROVIDERS=OPENID_CONNECT_PROVIDERS)
     @patch(('onadata.apps.api.viewsets.openid_connect_viewset.'