+++ CHANGES.rst
@@ -3,6 +3,51 @@ Changelog for Onadata
 
 ``* represents releases that introduce new migrations``
 
+v2.3.5(2020-06-18)
+------------------
+
+- Introduced caching for UserProfile objects
+  `PR #1823 <https://github.com/onaio/onadata/pull/1823>`_
+  [@WinnyTroy]
+- Send CRUD notifications for Forms, Submissions and SubmissionReviews
+  `PR #1793 <https://github.com/onaio/onadata/pull/1793>`_
+  [@lincmba]
+- Set enketo cookie `__enketo_meta_username` on login
+  `PR #1834 <https://github.com/onaio/onadata/pull/1834>`_
+  [@FrankApiyo]
+
+v2.3.4(2020-06-15)
+------------------
+
+- Use last name as first name if not present in OpenID Connect identification token
+  `PR #1831 <https://github.com/onaio/onadata/pull/1831>`_
+  [@DavisRayM]
+
+v2.3.3(2020-05-19)
+------------------
+
+- Fix an issue where file attachments/uploads were automatically soft-deleted
+  `PR #1816 <https://github.com/onaio/onadata/pull/1816>`_
+  [@DavisRayM]
+- Cache projects after creation and retrieve the project from cache in subsequent requests
+  `PR #1819 <https://github.com/onaio/onadata/pull/1819>`_
+  [@KipSigei]
+- Fix an issue where anonymous submissions from Enketo would fail
+  `PR #1825 <https://github.com/onaio/onadata/pull/1825>`_
+  [@WinnyTroy]
+- Add a management command that deletes users
+  `PR #1717 <https://github.com/onaio/onadata/pull/1717>`_
+  [@WinnyTroy]
+- Ensure that authenticated users can only submit to forms they have access to
+  `PR #1804 <https://https://github.com/onaio/onadata/pull/1804>`_
+  [@DavisRayM]
+- Add support for Tableau v2
+  `PR #1820 <https://github.com/onaio/onadata/pull/1820>`_
+  [@WinnyTroy]
+- Add setting to optional turn off creation of public projects & xforms
+  `PR #1829 <https://github.com/onaio/onadata/pull/1829>`_
+  [@DavisRayM]
+
 v2.3.2(2020-05-05)
 ------------------